Remove std::vector oppNode_ from MasterElement.
Also move Edge2D to its own source file.
